ajax 스니펫 추가

+++ src/main/java/shop/mtcoding/blog/service/BoardService.java
@@ -6,8 +6,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import shop.mtcoding.blog.dto.board.BoardReq.BoardSaveReqDto;
+import shop.mtcoding.blog.dto.board.BoardReq.BoardUpdateRespDto;
 import shop.mtcoding.blog.handler.ex.CustomApiException;
 import shop.mtcoding.blog.handler.ex.CustomException;
+import shop.mtcoding.blog.model.Board;
 import shop.mtcoding.blog.model.BoardRepository;
 
 @Transactional(readOnly = true)
@@ -27,11 +29,39 @@ public class BoardService {
     }
 
     @Transactional
-    public void 게시글삭제(int id) {
-        int result = boardRepository.deleteById(id);
-        if (result != 1) {
-            throw new CustomApiException("삭제 실패", HttpStatus.INTERNAL_SERVER_ERROR);
+    public void 게시글삭제(int id, int userId) {
+        Board boardPS = boardRepository.findById(id);
+        if (boardPS == null) {
+            throw new CustomApiException("없는 게시글을 삭제할 수 없습니다.");
+        }
+        if (boardPS.getUserId() != userId) {
+            throw new CustomApiException("해당 게시글을 삭제할 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+        try {
+            boardRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new CustomApiException("서버에 일시적인 문제가 생겼습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+            // 로그를 남겨야 함 (DB or File)
+
         }
     }
 
+    public void 게시글수정(int id, int userId, BoardUpdateRespDto boardUpdateRespDto) {
+        Board boardPS = boardRepository.findById(id);
+        if (boardPS == null) {
+            throw new CustomApiException("없는 게시글을 수정할 수 없습니다.");
+        }
+        if (boardPS.getUserId() != userId) {
+            throw new CustomApiException("해당 게시글을 수정할 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+        try {
+            boardRepository.updateById(boardUpdateRespDto.getId(), boardUpdateRespDto.getTitle(),
+                    boardUpdateRespDto.getContent());
+        } catch (Exception e) {
+            throw new CustomApiException("서버에 일시적인 문제가 생겼습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+            // 로그를 남겨야 함 (DB or File)
+
+        }
+
+    }
 }
